+++ packages/amplify-cli-core/src/errors/index.ts
@@ -15,3 +15,4 @@ export class AppNotFoundError extends Error {}
 export class AppAlreadyDeployedError extends Error {}
 export class SchemaDoesNotExistError extends Error {}
 export class AngularConfigNotFoundError extends Error {}
+export class AppIdMismatchError extends Error {}