+++ packages/amplify-provider-awscloudformation/src/configuration-manager.ts
@@ -9,7 +9,7 @@ import awsRegions from './aws-regions';
 import constants from './constants';
 import setupNewUser from './setup-new-user';
 import obfuscateUtil from './utility-obfuscate';
-import systemConfigManager from './system-config-manager';
+import * as systemConfigManager from './system-config-manager';
 import { doAdminTokensExist, getTempCredsWithAdminTokens, isAmplifyAdminApp } from './utils/admin-helpers';
 import { resolveAppId } from './utils/resolve-appId';
 import { AuthFlow, AuthFlowConfig, AwsSdkConfig } from './utils/auth-types';
@@ -303,7 +303,6 @@ async function setProjectConfigAction(context: $TSContext) {
     } else if (inputParams.configLevel === 'project') {
       context.exeInfo.awsConfigInfo.action = 'create';
       context.exeInfo.awsConfigInfo.configLevel = 'project';
-      context.exeInfo.awsConfigInfo.config = defaultAWSConfig;
     } else {
       context.exeInfo.awsConfigInfo.action = 'none';
       context.exeInfo.awsConfigInfo.configLevel = 'general';
@@ -319,7 +318,6 @@ async function setProjectConfigAction(context: $TSContext) {
       if (answer.setProjectLevelConfig) {
         context.exeInfo.awsConfigInfo.action = 'create';
         context.exeInfo.awsConfigInfo.configLevel = 'project';
-        context.exeInfo.awsConfigInfo.config = defaultAWSConfig;
       } else {
         context.exeInfo.awsConfigInfo.action = 'none';
         context.exeInfo.awsConfigInfo.configLevel = 'general';
@@ -351,6 +349,7 @@ async function promptForAuthConfig(context: $TSContext, authConfig?: AuthFlowCon
   if (availableProfiles && availableProfiles.length > 0) {
     let authType: AuthFlow;
     let isAdminApp = false;
+
     if (authConfig?.type) {
       authType = authConfig.type;
     } else {
@@ -362,6 +361,7 @@ async function promptForAuthConfig(context: $TSContext, authConfig?: AuthFlowCon
       }
       authType = await askAuthType(isAdminApp);
     }
+
     if (authType === 'profile') {
       printProfileInfo(context);
       awsConfigInfo.config.useProfile = true;
@@ -372,6 +372,9 @@ async function promptForAuthConfig(context: $TSContext, authConfig?: AuthFlowCon
       awsConfigInfo.configLevel = 'amplifyAdmin';
       awsConfigInfo.config.useProfile = false;
       return;
+    } else {
+      awsConfigInfo.config.useProfile = false;
+      delete awsConfigInfo.config.profileName;
     }
   } else {
     awsConfigInfo.config.useProfile = false;
@@ -579,10 +582,12 @@ export async function loadConfigurationForEnv(context: $TSContext, env: string, 
       exitOnNextTick(1);
     }
   } else if (authType.type === 'profile') {
-    if (authType?.profileName) {
+    try {
       awsConfig = await systemConfigManager.getProfiledAwsConfig(context, authType.profileName);
-    } else {
-      throw Error('Project configuration invalid. Missing profile name.');
+    } catch (e) {
+      context.print.error(`Failed to get profile: ${e.message || e}`);
+      await context.usageData.emitError(e);
+      exitOnNextTick(1);
     }
   } else if (authType.type === 'accessKeys') {
     awsConfig = loadConfigFromPath(projectConfigInfo.config.awsConfigFilePath);
@@ -703,7 +708,13 @@ export async function getAwsConfig(context: $TSContext): Promise<AwsConfig> {
   let awsConfig: AwsSdkConfig;
   if (awsConfigInfo.configLevel === 'project') {
     if (awsConfigInfo.config.useProfile) {
-      awsConfig = await systemConfigManager.getProfiledAwsConfig(context, awsConfigInfo.config.profileName);
+      try {
+        awsConfig = await systemConfigManager.getProfiledAwsConfig(context, awsConfigInfo.config.profileName);
+      } catch (e) {
+        context.print.error(`Failed to get profile: ${e.message || e}`);
+        await context.usageData.emitError(e);
+        exitOnNextTick(1);
+      }
     } else {
       awsConfig = {
         accessKeyId: awsConfigInfo.config.accessKeyId,