+++ packages/amplify-provider-awscloudformation/src/iterative-deployment/index.ts
@@ -1,1 +1,4 @@
-export { DeploymentManager } from './deployment-manager';
+export * from './deployment-manager';
+export * from './deployment-state-manager';
+export * from './constants';
+export * from './iterative-rollback';