+++ packages/amplify-category-predictions/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.8.2...amplify-category-predictions@2.8.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.8.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.8.1...amplify-category-predictions@2.8.2) (2021-04-14)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.8.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.7.6...amplify-category-predictions@2.8.1) (2021-04-09)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
 ## [2.7.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.7.5...amplify-category-predictions@2.7.6) (2021-03-23)